# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestionTriggered": {
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "IngestionTriggered",
        "type": "object"
      },
...
      "SourceCreate": {
        "additionalProperties": false,
        "properties": {
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "source_type": {
            "default": "rss",
            "enum": [
              "rss",
              "atom"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 32,
            "title": "Tags",
            "type": "array"
          },
          "url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url"
        ],
        "title": "SourceCreate",
        "type": "object"
      },
      "SourcePreferencePublic": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_muted": {
            "title": "Is Muted",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "preference_version": {
            "title": "Preference Version",
            "type": "integer"
          },
          "reputation_score": {
            "title": "Reputation Score",
            "type": "number"
          },
          "reputation_version": {
            "title": "Reputation Version",
            "type": "integer"
          },
          "url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "url",
          "origin",
          "is_active",
          "is_muted",
          "reputation_score",
          "reputation_version",
          "preference_version"
        ],
        "title": "SourcePreferencePublic",
        "type": "object"
      },
      "SourcePreferenceUpdate": {
        "additionalProperties": false,
        "properties": {
          "is_muted": {
            "title": "Is Muted",
            "type": "boolean"
          }
        },
        "required": [
          "is_muted"
        ],
        "title": "SourcePreferenceUpdate",
        "type": "object"
      },
      "SourcePublic": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_muted": {
            "title": "Is Muted",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "source_type": {
            "title": "Source Type",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "url",
          "source_type",
          "tags",
          "is_active",
          "origin",
          "is_muted"
        ],
        "title": "SourcePublic",
        "type": "object"
      },
      "SourceUpdate": {
        "additionalProperties": false,
        "properties": {
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source_type": {
            "anyOf": [
              {
                "enum": [
                  "rss",
                  "atom"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 32,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "title": "SourceUpdate",
        "type": "object"
      },
      "TelemetryDeletionPublic": {
        "properties": {
          "anonymized_audit_events": {
            "title": "Anonymized Audit Events",
            "type": "integer"
          },
          "deleted_usage_events": {
            "title": "Deleted Usage Events",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "deleted_usage_events",
          "anonymized_audit_events"
        ],
        "title": "TelemetryDeletionPublic",
        "type": "object"
      },
      "TelemetryExportPublic": {
        "properties": {
          "audit_count": {
            "title": "Audit Count",
            "type": "integer"
          },
          "audit_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Audit Events",
            "type": "array"
          },
          "usage_count": {
            "title": "Usage Count",
            "type": "integer"
          },
          "usage_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Usage Events",
            "type": "array"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "usage_count",
          "audit_count",
          "usage_events",
          "audit_events"
        ],
        "title": "TelemetryExportPublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Tuxnews",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/telemetry/{user_id}": {
      "delete": {
        "operationId": "delete_user_telemetry_route_api_v1_admin_telemetry__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryDeletionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete User Telemetry Route",
        "tags": [
          "telemetry"
        ]
      },
      "get": {
        "operationId": "export_user_telemetry_route_api_v1_admin_telemetry__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelemetryExportPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export User Telemetry Route",
        "tags": [
          "telemetry"
        ]
      }
    },
...
    "/api/v1/preferences/sources": {
      "get": {
        "operationId": "get_source_preferences_api_v1_preferences_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourcePreferencePublic"
                  },
                  "title": "Response Get Source Preferences Api V1 Preferences Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Source Preferences",
        "tags": [
          "preferences"
        ]
      }
    },
    "/api/v1/preferences/sources/{source_id}": {
      "patch": {
        "operationId": "edit_source_preference_api_v1_preferences_sources__source_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourcePreferenceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcePreferencePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Edit Source Preference",
        "tags": [
          "preferences"
        ]
      }
    },
...
    "/api/v1/sources": {
      "get": {
        "operationId": "list_sources_api_v1_sources_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourcePublic"
                  },
                  "title": "Response List Sources Api V1 Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sources",
        "tags": [
          "sources"
        ]
      },
      "post": {
        "operationId": "create_source_api_v1_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Source",
        "tags": [
          "sources"
        ]
      }
    },
    "/api/v1/sources/{source_id}": {
      "delete": {
        "operationId": "delete_source_api_v1_sources__source_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Source",
        "tags": [
          "sources"
        ]
      },
      "get": {
        "operationId": "get_source_api_v1_sources__source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Source",
        "tags": [
          "sources"
        ]
      },
      "patch": {
        "operationId": "update_source_api_v1_sources__source_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Source",
        "tags": [
          "sources"
        ]
      }
    },
    "/api/v1/sources/{source_id}/ingest": {
      "post": {
        "operationId": "trigger_source_ingestion_api_v1_sources__source_id__ingest_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionTriggered"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trigger Source Ingestion",
        "tags": [
          "sources"
        ]
      }
    },
...
  }
}